# docs/1.0.0-SNAPSHOT-api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: accounts
  description: accounts
- name: statements
  description: statements
paths:
  /statements:
    get:
      tags:
      - statements
      - Statements
      summary: Get Statements
      operationId: getStatements
      parameters:
      - name: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        in: header
        required: true
        schema:
          type: string
      - name: "The time when the PSU last logged in with the TPP.  All dates in the\
          \ HTTP headers are represented as RFC 7231 Full Dates. An example is below:\
          \  Sun, 10 Sep 2017 19:43:31 UTC"
        in: header
        required: false
        schema:
          type: string
      - name: The PSU's IP address if the PSU is currently logged in with the TPP.
        in: header
        required: false
        schema:
          type: string
      - name: An RFC4122 UID used as a correlation id.
        in: header
        required: false
        schema:
          type: string
      - name: Indicates the user-agent that the PSU is using.
        in: header
        required: false
        schema:
          type: string
      - name: "The UTC ISO 8601 Date Time to filter statements FROM NB Time component\
          \ is optional - set to 00:00:00 for just Date. If the Date Time contains\
          \ a timezone, the ASPSP must ignore the timezone component."
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: "The UTC ISO 8601 Date Time to filter statements TO NB Time component\
          \ is optional - set to 00:00:00 for just Date. If the Date Time contains\
          \ a timezone, the ASPSP must ignore the timezone component."
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/json:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/jose+jwe:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
  /accounts/{AccountId}/statements:
    get:
      tags:
      - Statements
      - accounts
      summary: Get Statements
      operationId: getAccountsAccountIdStatements
      parameters:
      - name: AccountId
        in: path
        required: true
        schema:
          type: string
      - name: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        in: header
        required: true
        schema:
          type: string
      - name: "The time when the PSU last logged in with the TPP.  All dates in the\
          \ HTTP headers are represented as RFC 7231 Full Dates. An example is below:\
          \  Sun, 10 Sep 2017 19:43:31 UTC"
        in: header
        required: false
        schema:
          type: string
      - name: The PSU's IP address if the PSU is currently logged in with the TPP.
        in: header
        required: false
        schema:
          type: string
      - name: An RFC4122 UID used as a correlation id.
        in: header
        required: false
        schema:
          type: string
      - name: Indicates the user-agent that the PSU is using.
        in: header
        required: false
        schema:
          type: string
      - name: "The UTC ISO 8601 Date Time to filter statements FROM NB Time component\
          \ is optional - set to 00:00:00 for just Date. If the Date Time contains\
          \ a timezone, the ASPSP must ignore the timezone component."
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: "The UTC ISO 8601 Date Time to filter statements TO NB Time component\
          \ is optional - set to 00:00:00 for just Date. If the Date Time contains\
          \ a timezone, the ASPSP must ignore the timezone component."
        in: query
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/json:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/jose+jwe:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
  /accounts/{AccountId}/statements/{StatementId}:
    get:
      tags:
      - Statements
      - accounts
      summary: Get Statements
      operationId: getAccountsAccountIdStatementsStatementId
      parameters:
      - name: StatementId
        in: path
        required: true
        schema:
          type: string
      - name: AccountId
        in: path
        required: true
        schema:
          type: string
      - name: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        in: header
        required: true
        schema:
          type: string
      - name: "The time when the PSU last logged in with the TPP.  All dates in the\
          \ HTTP headers are represented as RFC 7231 Full Dates. An example is below:\
          \  Sun, 10 Sep 2017 19:43:31 UTC"
        in: header
        required: false
        schema:
          type: string
      - name: The PSU's IP address if the PSU is currently logged in with the TPP.
        in: header
        required: false
        schema:
          type: string
      - name: An RFC4122 UID used as a correlation id.
        in: header
        required: false
        schema:
          type: string
      - name: Indicates the user-agent that the PSU is using.
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/json:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/jose+jwe:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadStatement2"
  /accounts/{AccountId}/statements/{StatementId}/transactions:
    get:
      tags:
      - Transactions
      - accounts
      summary: Get Transactions
      operationId: getAccountsAccountIdStatementsStatementIdTransactions
      parameters:
      - name: StatementId
        in: path
        required: true
        schema:
          type: string
      - name: AccountId
        in: path
        required: true
        schema:
          type: string
      - name: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        in: header
        required: true
        schema:
          type: string
      - name: "The time when the PSU last logged in with the TPP.  All dates in the\
          \ HTTP headers are represented as RFC 7231 Full Dates. An example is below:\
          \  Sun, 10 Sep 2017 19:43:31 UTC"
        in: header
        required: false
        schema:
          type: string
      - name: The PSU's IP address if the PSU is currently logged in with the TPP.
        in: header
        required: false
        schema:
          type: string
      - name: An RFC4122 UID used as a correlation id.
        in: header
        required: false
        schema:
          type: string
      - name: Indicates the user-agent that the PSU is using.
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadTransaction6"
            application/json:
              schema:
                $ref: "#/components/schemas/OBReadTransaction6"
            application/jose+jwe:
              schema:
                $ref: "#/components/schemas/OBReadTransaction6"
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/OBReadTransaction6"
  /accounts/{AccountId}/statements/{StatementId}/file:
    get:
      tags:
      - Statements
      - accounts
      summary: Get Statements
      operationId: getAccountsAccountIdStatementsStatementIdFile
      parameters:
      - name: StatementId
        in: path
        required: true
        schema:
          type: string
      - name: AccountId
        in: path
        required: true
        schema:
          type: string
      - name: An Authorisation Token as per https://tools.ietf.org/html/rfc6750
        in: header
        required: true
        schema:
          type: string
      - name: "The time when the PSU last logged in with the TPP.  All dates in the\
          \ HTTP headers are represented as RFC 7231 Full Dates. An example is below:\
          \  Sun, 10 Sep 2017 19:43:31 UTC"
        in: header
        required: false
        schema:
          type: string
      - name: The PSU's IP address if the PSU is currently logged in with the TPP.
        in: header
        required: false
        schema:
          type: string
      - name: An RFC4122 UID used as a correlation id.
        in: header
        required: false
        schema:
          type: string
      - name: Indicates the user-agent that the PSU is using.
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json;charset=utf-8:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
                format: binary
            application/jose+jwe:
              schema:
                type: string
                format: binary
            application/json; charset=utf-8:
              schema:
                type: string
                format: binary
components:
  schemas:
    OBReadStatement2:
      type: object
    OBReadTransaction6:
      type: object