# src/main/java/com/openbanking/statements/model/OBReadStatement2.java
package com.openbanking.statements.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadStatement2
 */
@Getter
@Setter
public class OBReadStatement2 implements Serializable {
	private static final long serialVersionUID = 1L;

}