# src/main/java/com/openbanking/statements/model/OBReadTransaction6.java
package com.openbanking.statements.model;

import java.io.Serializable;
import lombok.Getter;
import lombok.Setter;

/**
 * OBReadTransaction6
 */
@Getter
@Setter
public class OBReadTransaction6 implements Serializable {
	private static final long serialVersionUID = 1L;

}